requirements.txt
# №1: Список Python-зависимостей
aiogram==3.13.1
aiosqlite==0.20.0
python-dotenv==1.0.1